fix: reconcile existing sidecar config, type PortInferenceServerInternal as int32
- When sidecar already exists, reconcile BACKEND_URL, image, and ports
  to prevent stale config from breaking decode routing
- Declare PortInferenceServerInternal as int32 for consistency with
  PortInferenceServer, removing unnecessary casts

+++ pkg/utils/consts/consts.go
@@ -114,7 +114,7 @@ const (
 	RoutingSidecarTag   = "v0.7.1"
 
 	// When the routing sidecar is present, vLLM moves to this internal port.
-	PortInferenceServerInternal = 5001
+	PortInferenceServerInternal = int32(5001)
 
 	// Inference role constants for P/D disaggregated serving.
 	InferenceRoleEnvName = "KAITO_INFERENCE_ROLE"
